+++ luluisher/carts/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from carts.views import CartView
+
+urlpatterns = [
+    path('', CartView.as_view())
+]
