+++ luluisher/luluisher/settings.py
@@ -45,7 +45,8 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'corsheaders',
     'users',
-    'products'
+    'products',
+    'carts'
 ]
 
 MIDDLEWARE = [
