+++ luluisher/luluisher/urls.py
@@ -2,6 +2,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('users', include('users.urls')),
-    path('products', include('products.urls') )
+    path('carts', include('carts.urls')),
 
 ]
