+++ luluisher/users/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
 
-from users.views import SignUpView, SignInView, CartView
+from users.views import SignUpView, SignInView
 
 urlpatterns = [
     path('/signup', SignUpView.as_view()),
     path('/signin', SignInView.as_view()),
-    path('/cart', CartView.as_view())
 ]
